services/user-profile-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
sqlalchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.1
httpx==0.27.0